collect incentives switch with spread rewards and pseudo logic with todos for gauges duration uptime
