+++ smart-contracts/contracts/cl-vault/src/rewards/mod.rs
@@ -1,3 +1,4 @@
+pub mod autocompound;
 mod distribution;
 mod helpers;
